// backend/src/cad-conversion/cad-conversion.service.d.ts
import { PrismaService } from '../prisma/prisma.service';
export declare class CadConversionService {
    private prisma;
    constructor(prisma: PrismaService);
    create(data: any): Promise<{
        status: string;
        id: number;
        createdAt: Date;
        updatedAt: Date;
        fileSize: number | null;
        originalFileName: string;
        originalFilePath: string;
        svgFilePath: string | null;
        cadVersion: string | null;
        conversionTime: number | null;
        entityCount: number | null;
        layerCount: number | null;
        blockCount: number | null;
        errorMessage: string | null;
    }>;
    findAll(): Promise<{
        status: string;
        id: number;
        createdAt: Date;
        updatedAt: Date;
        fileSize: number | null;
        originalFileName: string;
        originalFilePath: string;
        svgFilePath: string | null;
        cadVersion: string | null;
        conversionTime: number | null;
        entityCount: number | null;
        layerCount: number | null;
        blockCount: number | null;
        errorMessage: string | null;
    }[]>;
    findOne(id: number): Promise<{
        status: string;
        id: number;
        createdAt: Date;
        updatedAt: Date;
        fileSize: number | null;
        originalFileName: string;
        originalFilePath: string;
        svgFilePath: string | null;
        cadVersion: string | null;
        conversionTime: number | null;
        entityCount: number | null;
        layerCount: number | null;
        blockCount: number | null;
        errorMessage: string | null;
    }>;
    update(id: number, data: any): Promise<{
        status: string;
        id: number;
        createdAt: Date;
        updatedAt: Date;
        fileSize: number | null;
        originalFileName: string;
        originalFilePath: string;
        svgFilePath: string | null;
        cadVersion: string | null;
        conversionTime: number | null;
        entityCount: number | null;
        layerCount: number | null;
        blockCount: number | null;
        errorMessage: string | null;
    }>;
    remove(id: number): Promise<{
        status: string;
        id: number;
        createdAt: Date;
        updatedAt: Date;
        fileSize: number | null;
        originalFileName: string;
        originalFilePath: string;
        svgFilePath: string | null;
        cadVersion: string | null;
        conversionTime: number | null;
        entityCount: number | null;
        layerCount: number | null;
        blockCount: number | null;
        errorMessage: string | null;
    }>;
}